ml/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
transformers
torch
numpy
